Begin work on showing admins and admin groups that admins can view.
When limited admins have view privileges and they are viewing groups,
the groups were returning all admins that belong to those groups.
However, if viewing the admin show/index pages, some of those admins
would be missing, since the limited admin could only view admins they
have full privileges to edit (so they need to be authorized to all of
the groups that admin belongs to). While this is the correct permissions
for editing admins, we want to loosen the ability to view/show those
admins, so the admin listings can be more complete.

At the same time, when viewing admins, we then need to ensure we only
return groups that the limited admin should know about, and not
necessarily all of the groups that the admin belongs to (since they
might be groups outside of the limited admin's privileges).