Bring back Trafficserver DNS configuration to try and improve resilience
Now that Envoy may be a remote server, it's possible it requires DNS
lookups to connect. So bring back some of our older configuration to
explicitly configure the DNS and keepalive settings, but with hard-coded
long timeouts to try and ensure any DNS outages don't affect the
Trafficserver to Envoy communication hop.