Fix rate limiting bug when anonymous usage is used and overlapping IPs
If an API backend is configured to allow for anonymous access without
API keys, then there are some options to allow the authenticated rate
limits only use the API key limits and not the IP limits. However, due
to some caching of when an IP exceeds rate limits, this could still lead
to a user with an API key being blocked once the IP exceeded the
configured anonymous limit for the IP.

The bug was due to the caching mechanism to check for cached rate limit
exceedings looping over all possible rate limits, and not excluding
these limits that we don't actually increment. The increment logic was
properly skipping over the non-applicable rate limits, so we just needed
to shift where the check for non-applicable limits was being applied so
it got applied to all loops when checking for rate limits, not just in
certain areas.

As part of this, this also refactors some of the rate limit test helpers
to use keywoard arguments so it's a little more obvious what methods use
what arguments, rather than a big hash of optional values being passed
around. The need to filter these makes this perhaps still a little
clunky, but I was having a hard time tracing things originally, so
hopefully this will make it easier.