Add db statement_timeout and allow for longer running analytics queries
Setup PostgreSQL `statement_timeout` for generally better safety to
prevent long running queries.

But for the analytics summary endpoint, specifically allow long running
queries. Also shift the analytics summary endpoint so it should be able
to complete in the background if it's not cached yet, rather than being
subject to any request timeouts.