Implement cache for geoip download.
If we hit the download rate limit due to apps being redeployed, then
that can lead to geoip data going missing until the download limits
expire and things can be redownloaded. This is more likely to happen as
the number of concurrent replicas increases, so shift all of this to an
external process than can just run once a day, and then all of our
workers can download from the cache.

This could eventually be cleaned up and better integrated, but just
trying to do something simple to implement for now.